Broke out AverageResidue into its specific classes.
