+++ mzLib/MassSpectrometry/Deconvolution/AverageResidue/AverageResidue.cs
@@ -1,0 +1,12 @@
+﻿namespace MassSpectrometry;
+
+public abstract class AverageResidue
+{
+    protected const double fineRes = 0.125;
+    protected const double minRes = 1e-8;
+    protected static readonly int NumAveraginesToGenerate = 1500;
+    public abstract int GetMostIntenseMassIndex(double testMass);
+    public abstract double[] GetAllTheoreticalMasses(int index);
+    public abstract double[] GetAllTheoreticalIntensities(int index);
+    public abstract double GetDiffToMonoisotopic(int index);
+}
